Code cleanup and restructuring
