+++ components/modules/MentorCard/variants/Mobile.tsx
@@ -1,0 +1,4 @@
+import { MentorCardProps } from "./Base";
+export function Mobile(props: MentorCardProps) {
+    return <div></div>;
+}
