+++ pages/index.tsx
@@ -3,7 +3,7 @@ import Head from "next/head";
 import Link from "next/link";
 import { Button, Container, Heading, Text } from "@chakra-ui/react";
 
-import { MentorCard, MentorCardProps } from "components/modules/MentorCard/MentorCard";
+import { MentorCard, MentorCardProps } from "components/modules/MentorCard";
 
 const mentor: Omit<MentorCardProps, "variant"> = {
     fullname: "Bjarne Stroustrup",
@@ -12,10 +12,19 @@ const mentor: Omit<MentorCardProps, "variant"> = {
     location: "New York, NY",
     expInYears: 40,
     stories: "Too busy saving the world from evil coders",
-    tags: ["software engineer", "c++"],
+    tags: [
+        "software engineer",
+        "c++ creator",
+        "c++23",
+        "c++20",
+        "c++17",
+        "c++14",
+        "c++11",
+        "c++98",
+    ],
     avgRating: 4.3,
-    numberOfEndorsements: 93,
-    numberOfReviews: 127,
+    noEndorsements: 93,
+    noReviews: 127,
 
     short:
         "In 1979, Stroustrup began his career as a member of technical staff in the Computer Science Research Center of Bell Labs in Murray Hill, New Jersey, USA. There, he began his work on C++ and programming techniques",
