+++ tailwind.config.js
@@ -4,6 +4,7 @@ const colors = {
     ...colorPalette,
     primary: colorPalette.coolGray[900],
     secondary: colorPalette.coolGray[500],
+    fade: "#DDE2F2",
 };
 
 module.exports = {
@@ -28,6 +29,20 @@ module.exports = {
 
             xl: "1280px",
             // => @media (min-width: 1280px) { ... }
+        },
+        fontSize: {
+            xs: ".75rem",
+            sm: ".875rem",
+            md: "0.875rem",
+            base: "0.875rem",
+            lg: "1.125rem",
+            xl: "1.25rem",
+            "2xl": "1.5rem",
+            "3xl": "1.875rem",
+            "4xl": "2.25rem",
+            "5xl": "3rem",
+            "6xl": "4rem",
+            "7xl": "5rem",
         },
     },
     variants: {
